Remove FileSource
